version 1.0.5 , explode middleware , update README.md , update keywords package.json
